Automatically detect and set correct panel cmdline
This script will find the value given by the kernel with get_panel_maker_id

	LCD_RENESAS_LGD = 0,
	LCD_RENESAS_JDI = 1

The lcd_maker_id value is generated from the panel hardware, is always accurate and will always set the correct cmdline of your panel

This script can also simply be copy pasted for a kernel zip in loki.sh before the loki function. It has no other requirements besides the script itself since it directly searches and modifies the boot.img and doesn't require the boot.img to be unpacked.

Do not shorten the search function as we need to ensure only the cmdline is being changed and nothing in the actual kernel source.

Continue building the kernel with either the full lgd or jdi cmdline, the script will change the panel as needed.

Currently the script looks for the extracted boot.img in /tmp and /tmp/loki

Change-Id: I37a31e074d33b2b93bc6f52efbadce2986606e3d
Signed-off-by: dr87 <nb00870@gmail.com>